pyproject.toml
###########
# 📜 Poetry
###########
[tool.poetry]
name = "fairness_tales_workshop"
version = "0.1.0"
description = "Fairness Tales: How To Measure And Mitigate Unfair Bias in Machine Learning Models"
authors = ["John Sandall <contact@coefficient.ai>"]
license = "UNLICENSED"
classifiers = ["Private :: Do Not Upload"]
packages = [
  { include = "fairness_tales_workshop", from = "fairness_tales_workshop" },
]

[tool.poetry.dependencies]
python = "^3.12"
# Everything below here is alphabetically sorted
fairlearn = "^0.11.0"
joblib = "^1.4.2"
jupyter-black = "^0.4.0"
jupyterlab = "^4.3.1"
lightgbm = "^4.5.0"
matplotlib = "^3.9.3"
numpy = "^2.1.3"
openai = "^1.56.0"
pandas = "^2.2.3"
pickleshare = "^0.7.5"
pyarrow = "^18.1.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
tabulate = "^0.9.0"
xgboost = "^2.1.3"
jupytext = "^1.16.4"

[tool.poetry.dev-dependencies]
# Everything below here is alphabetically sorted
ipdb = "^0.13.13"
pip-audit = "^2.7.3"
pre-commit = "^3.8.0"
pytest = "^7.4.4"
ruff = "^0.1.14"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

############
# ✅ Linters
############
[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = [
  "A",
  "AIR",
  "ARG",
  "B",
  "C4",
  "C90",
  "COM",
  # "CPY",
  "D",
  "DJ",
  "DTZ",
  "E",
  # "E4", "E7", "E9",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FIX",
  "FLY",
  # "FURB",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  # "LOG",
  "N",
  "NPY",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TCH",
  "TD",
  "TID",
  "TRY",
  "UP",
  "W",
]
fixable = ["ALL"]
ignore = [
  "ARG001",
  "B018",
  "D100",
  "D103",
  "E501",
  "ERA001",
  "INP001",
  "NPY002",
  "PD901",
  "PERF401",
  "PTH123",
  "RET503",
  "SIM108",
  "T201",
  "TRY201",
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
